Added powerup selection and info
The selection scene now lets you select which powerups should be active for any particular fight. This is only for testing, as powerups will incrementally unlock and stay unlocked during regular play.
Also added a window to display the details of each powerup type, also in the selection scene.
Also also added the beginning of the sword powerup.