Added the Interact Overlay
In the overworld, moving to an enemy space now opens up the interact overlay, which contains buttons for Battle, Talk, Info, and Back. Battle starts a battle with the enemy, Info displays info about how the enemy is susceptible to different powerups, and Back returns to the overworld. Talk clears out spaces for dialogue, but that hasn't been fully implemented yet.