Beginnings of dialogue system
Each OverworldEnemySpace now has a DialogueStep list, which is serializable in the inspector. Each DialogueStep contains data about who is talking and what they are saying, with potential future options to have specific custom event triggers included in the dialogue as well, such as playing animations and updating game states.
Clicking the Talk button now shows the first DialogueStep, and clicking the Back button (renamed "Next") advances the dialogue to the next stage.