Cutscene improvements pt 7
Cutscene backgrounds are now filled from a prefab, rather than a static image.

Animations for specific characters in a cutscene can be started, by using the PlayAnimation cutscene type. The gameobject name for the character that should be animated, as well as the name of the animation, can be provided as strings for the cutscene step.