Added tutorial 3
Tutorial 3 introduces the healing and water powerups, and concludes the battles for the tutorial overworld.
Added the ability to load a puzzle with a predefined arrangement of powerups.
Added the SubmitHealingWord, PlayerGetsHealed, SubmitWaterWord, WaterFillsPage, and WaterDrains tutorial advancement conditions.
Reworked the tutorial overworld to have only 3 tutorial battles. More beautification work is required here.