Overhauled the cutscene system
Cutscenes are now handled similarly to tutorials, with a series of pre-defined steps listed out in order within the cutscene manager file. Which cutscene is used is dependent upon what value StaticVariables.cutsceneID has, which is provided by the Overworld Scene before the cutscene scene is loaded.
Ported over the entirety of the opening cutscene into the new system, with some improvements throughout.
Added a rain effect that happens when the book is floating in the air atop the water spray that comes out of the well. Improved some other timings, dialogue lines, and chatheads as well.