Opening cutscene improvements
The switch statement that was used to implement each separate cutscene step has been replaced with a set of if statements using an incrementing index value.
All 3 backgrounds for the opening cutscene have actual assets in them. This involved a lot of repositioning of characters, and creation of some new assets like the fire animated prefabs.