Intro and general overworld improvements
Improved the appearance of the path between enemies and the spot that the player stands on.
A new book icon represents cutscenes in the overworld.
Made the intro overworld look nicer, with flowers etc.
The enemy name is now shown in the interact overlay, and the talk and info buttons have been moved to the same line to make room for it.
Added an icon to the info button that shows when there is info available for the enemy.
If an overworld space is a tutorial space, then the info button is not shown in the interact overlay for that encounter. Instead, the talk button is made larger.