Misc improvements
Number of powerups per puzzle and which powerups are active are now set by the overworld scene manager.

The icon that shows a powerup is available for a letter is now a colored gem, instead of a colored circle.

The magnitude of a powerup is now indicated by the number of powerup icons shown above the word strength.

The word display text on the submit word button now can scale in size for larger words, which has made room for the countdown and word strength sections to be made larger.