Improved Map Scene in prep for Alpha demo
The map scene now has world selections for the hometown and grasslands, with artwork of each area shown.
Below the world selection section, there is a debug settings area, which lets you change game completion progress, toggle the new auto submit feature, and enter a battle against a beholder with all powerups available.

Cleaned up the general scene manager script to remove some of the temporary junk that was there for previous demos.

Added the auto submit feature, which will clear or submit words automatically as soon as you lift your finger off the puzzle.

Also improved the pause button to have an actual pause symbol, and to animate the bookmark sliding off/on screen when pausing is toggled.
In the overworld, the "return to map" button now has a little home symbol on it.