Added Info Icon to DialogueManager
The dialogue manager has its own version of the info icon from the interaction manager. This one is only used for the fake buttons sliding animation to transition from the interaction manager to the dialogue manager. The dialogue manager version is synchronized to match the animation of the already visible version from the interaction manager.