Major rewrite to how world and stage progress is stored
Added the LaunchGameSetup scene, which has a list of all enemies in all worlds and generates a linked list of every stage in the game. The LaunchGameSetup code is run when the game opens, then immediately loads the normal Homepage scene.

Created the StageData object in the StaticVariables script, which holds information pertaining to a particular stage (world number and name, stage number within that world, the prefab for the enemy, and links to the previous and next stages).
All code referencing world "levels" (now stages) and world numbers by separate values have now been replaced with a reference to the single StageData object for that stage.
Hometown is now World 1, and Grasslands is now World 2.
A lot of the code has been simplified and made more readible with these changes.