Overworld border removal part 1
Work in progress removing the border on the overworld scenes, and making them expand to fill various screen sizes.
Hometown is complete, but grasslands still has a ways to go.
Created Scene Window 1.prefab and Scene Window 2.prefab. Number 2 is what will eventually become the main one when the update is finished, and the other 2 versions (#1 and the original) will be removed.