Fixed Interact Overlay height adjustment on various screen sizes
When the interact overlay moves or changes size, the overworld view will now move to keep the player at the right height if necessary.
As part of this change, the height that the player must be kept above the interact overlay is now based on a gameobject attached to the player sprite, rather than being based off of a fixed pixel amount.