Improved homepage design and layout
The continue adventure button now looks like a differently-colored book for each world. The books are recolors of the magic book used in the game, and there is one for each color of powerup. These books will possibly be used for selecting powerup specialization before battles in the future.
The overworld previews for the continue adventure option now uses higher resolution game screenshots.
Changed the book design for endless mode, and added more scattered pages behind the Settings page.
Added drop shadows for each book on the homepage.