Player position in the hierarchy now changes as they move through the overworld
As the player moves through the overworld, their position in the hierarchy changes so that they are always directly above the path / enemy space spots. This gives more flexibility in background design and allows for simulated depth as the player moves.
Also adjusted the position of some background elements in the hierarchy in the desert scene.