Fixed some win/loss issues
Winning or losing now correctly reveals the Victory/Defeat page as the page turn animation progresses.
The victory and defeat pages have been combined into one singular gameobject, with text that changes depending on the game end condition.
Created a TEST ENEMY enemy with dummy victory dialogue. This is a hack to fix an issue that prevents the player from continuing if they defeat an enemy with no victory dialogue.