Introduced the Read button to the Interact Overlay
The read button does not have any functionality yet.
The info and read buttons are now both always shown to the player, but are greyed-out, locked, and unusable until the player gets far enough in the story.
The info button becomes available starting with the ogre in the grasslands, and the read button becomes available starting with the paladin in the desert.
Improved how the Dialogue Manager is transitioned to and from the Interact Overlay, including a new way that the current Interact Overlay state is mimiced during the transitions.