Finished implementing the buffed type feature
Once reaching world 4, the player is given the ability to select a buffed powerup type, which makes it much more common when generating puzzles, and guarantees at least one buff of that type appears on each puzzle.
Modified the Interact Overlay to allow the player to pick their chosen buffed type out of all of the types unlocked.
I swear it took a lot of coding work to get this feature implemented and looking nice, you just gotta trust me on this one bruv.