Work in progress improving emotions
Added images that will eventually go on top of characters when they are speaking to show emotions. Included are angry, defeated, happy, questioning, worried, and surprised, although defeated might not need additional artwork at the end of the day. Excited will likely be removed as an emotion option, and normal will not have any additional artwork required.
Added a new version of the Player Angry chathead, and the brand new Player surprised chathead.