Improved the emotion flairs, added Surprise emotion
Emotion flair is the new term for the additional sprites that appear over the player chathead to convey most emotions.  Currently only the player has these emotion flairs, but the possibility exists for them to be added to enemies in the future.
Also added Surprised to the list of possible emotions.