Dialogue and cutscene improvement for grasslands
All of the dialogue for the grasslands is now better, except for the cave cutscene.
Added a new in-dialogue animation for magical glows, which is now used when healing groats after the battle with the goblin generals.
Fixed an issue where the final cutscene for an area was not clickable in the overworld.
Added powerup colors to the LaunchGameSetup script, where they can be edited from the inspector. The colors are saved to StaticVariables, and can be referenced by any script. This allows a lot of scripts to stop holding their own individual references to powerup colors in the future.