More work to add unique backgrounds for each powerup type
The unique backgrounds now appear when forming a word.
Resized the fonts for letter and word displays to be less blurry.
Letter spaces now support a unique "gem" image for each powerup type. Plain letters are black, but when there is a powerup available, or when part of a word, the letter turns white with a black border.