Fairy and Wizard attack improvements
The fairy now starts at 0 stacks, and builds up to a maximum of 5. Using the power of healing effectively removes 2 stacks now.
The wizard fire is now cured by making an attack while the book is flooded.
Roger now is wearing purple clothes.
Repositioned characters in the intro cutscene to be more out of the way. When the townsfolk are fighting the goblins, both characters are doing their walk animations on loop.
Letter space animations are now correctly synchronized once again.