Player Attack refactor
Created the AttackData object, which contains all the relevant data for a single player attack. This object gets created by the BattleManager. When an attack is started, the attackData is added to a queue of in-progress attacks. When the player is free to start an attack, an attack is removed from the queue and passed to the player attack animator, which starts animations and sends a projectile. The same attack data object is passed all the way to the point when the attack lands, at which point the data is parsed and the attack effect is applied. The advantage of this new system is that the entire battle state is captured when the word is submitted, and information like the page being flooded is all tracked inside the attack data itself.
Removed the pebble mechanic, as it will be replaced by a new earth effect shortly.
Powerup level is no longer a thing, so all effects that number of powerups have on an attack effect have been removed.