Added a bunch of desert stuff
Added the desert intro cutscene, as well as dialogue for the knight and skeleton.
The knight serves as the fire tutorial, but the tutorial is mostly unfinished at the moment.
The skeleton has been made the last enemy of the upcoming playtest, so it's been given some explanatory cheeky dialogue, 0 damage potential, and a whopping 5000 hp. Just to give the players something juicy to play around with once they complete the playtest.