Many small improvements
Shortened the height of the water spray during the intro cutscene to make room for the back button.
Added a custom letters layout to the fire tutorial.
Adjusted the battle background in the hometown dog fight, one element was only partially in view.
Adjusted the desert background on the homepage so the enemy sprite is in a more reasonable spot.
The flooding animation now disappears when the enemy's health hits 0.
Added a new type of tutorial advance condition, EnemyTakesDamageWithEmptyQueue, used when the player makes multiple attacks in sequence to ensure that all of them go off before proceeding.
During the upcoming playtest, the player can no longer choose a world later than the desert.