A bunch more misc changes
Updated the app icon and splash screen logo.
Changed npc movements in the tavern cutscene.
Cutscenes at the end of a world will now automatically load into the next world, instead of going back to the last overworld first.
Bumped up all enemy healths to account for the power creep of powerups.
Added some more dialogue from the player and children when the town is on fire during the intro cutscene.