More small tweaks
During tutorials, the earth buff bottom segment now starts as a child of the dialoguemanager, and then is moved next to the other earth buff parts, to guarantee it is just behind the top of the dialoguemanager no matter what the aspect ratio is.
The earth buff is no longer applied if the player or enemy health is 0.
Disabled the feature that changed the attack strength color if the page is flooded.
Minor dialogue text changes.