Fixed (hopefully) screen black issue
Many scenes did not manually start the generalscenemanager screen fade function, which occasionally caused some load order issues. if the generalscenemanager start function happened before the overworldscenemanager start function, you'd get a black screen.
Also, now during a battle, the game checks every frame to see if the player can start an attack. This removes a lot of the messy timings and conditional checks. Now a word can be added to the queue and it'll automatically be used as soon as available.