Created a python script for updating the word lists
Added list_generator.py, which can be run to modify the word lists used in the game. By adding words to the 'input - words to add' or 'input - words to remove' files, the 'list of all words' file will be updated when the script is run.
Adding words to the 'input - slurs' list will remove them from the list of all words entirely.
Adding words to the 'input - swear words' list will create a copy of the list of all words, but with the swears removed.
An impending code change will add a profanity toggle, which will determine if the swear or non-swear lists are used in puzzle generation and word validation.
Then, the swear and non-swear word lists are split by letter count and saved individually.