A bunch of misc changes
Removed some non-words from the word lists.
Renamed the pharaoh.
Added a "put away book, flipped" animation for the forest intro.
Rephrased the necromancy info description.
The manticore will now stop the "Lightning 2" attack if the electricity is dissipated while the attack bar is refilling.
The last step of dialogue and cutscenes now has the button labeled "END".
Fixed a bug allowing the player to have an invalid book selection after changing to a previous world in the settings.
Fixed a missing chathead in the mummy victory dialogue.
Horde enemies will be able to attack once again.
Rephrased the pharaoh's dialogue.